Add size() and clear() methods
public int size() method returns the number of nodes including the root
in the Binary Tree rooted at the node on which this method was called.

public void clear() method removes all of the mappings from the node on
which this method was called. It resets the BinaryTree with null values.